ProcessJsonMigration do not fetch all scenario JSONs at once
